Release 0.1.5 into master

+++ kairosActivity/src/test/java/com/kairos/activity/service/activity/PlannedTimeTypeIntegrationTest.java
@@ -3,7 +3,7 @@ package com.kairos.activity.service.activity;
 import com.kairos.activity.KairosActivityApplication;
 import com.kairos.activity.client.dto.RestTemplateResponseEnvelope;
 import com.kairos.activity.service.exception.ExceptionService;
-import com.kairos.response.dto.web.wta.PresenceTypeDTO;
+import com.kairos.response.dto.web.presence_type.PresenceTypeDTO;
 import org.junit.Assert;
 import org.junit.FixMethodOrder;
 import org.junit.Test;
@@ -18,9 +18,9 @@ import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.junit4.SpringRunner;
-import org.springframework.web.client.RestTemplate;
 
 import javax.inject.Inject;
+import java.math.BigInteger;
 import java.util.Date;
 import java.util.List;
 
@@ -51,7 +51,7 @@ public class PlannedTimeTypeIntegrationTest {
         ResponseEntity<RestTemplateResponseEnvelope<PresenceTypeDTO>> response = testRestTemplate.exchange(baseUrl, HttpMethod.POST, requestEntity, typeReference);
         System.out.println("reponse: "+response.getStatusCode());
         Assert.assertEquals(HttpStatus.CREATED, response.getStatusCode());
-        createdId = response.getBody().getData().getId();
+        createdId = response.getBody().getData().getId().longValue();
     }
 
     @Test
@@ -68,7 +68,7 @@ public class PlannedTimeTypeIntegrationTest {
     public void case3_updatePlannedTimeType() throws Exception{
         String baseUrl = getBaseUrl(24L, 4L)+"/"+createdId;
         System.out.println("baseUrl: "+baseUrl);
-        PresenceTypeDTO presenceTypeDTO = new PresenceTypeDTO("PlannedTimeType-updated"+createdId, createdId);
+        PresenceTypeDTO presenceTypeDTO = new PresenceTypeDTO("PlannedTimeType-updated"+createdId, BigInteger.valueOf(createdId));
         HttpEntity<PresenceTypeDTO> requestEntity = new HttpEntity<>(presenceTypeDTO);
         ParameterizedTypeReference<RestTemplateResponseEnvelope<PresenceTypeDTO>> responseRef = new ParameterizedTypeReference<RestTemplateResponseEnvelope<PresenceTypeDTO>>() {
         };
